move implementação específica de client opensearch para raiz
